♻️ Organize controllers into Account, Settings, and Api subdirectories
Group the flat controller directory (30 files) into logical subdirectories:

- Account/ (5): User self-service controllers (account, alias, openpgp, 2fa, voucher)
- Settings/ (11): Admin management controllers (settings, domains, api tokens, etc.)
- Api/ (7): External integration endpoints (dovecot, postfix, keycloak, wkd, mta-sts, etc.)
- Root (7): Auth, registration, recovery, init, and Sonata CRUD controllers

VoucherSettingsController is renamed to Settings\VoucherController since the
namespace collision with Account\VoucherController no longer exists.

Route names and URL paths are unchanged, so templates, Behat features,
security config, and all other references remain unaffected.
