+++ src/Controller/Account/AliasController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Controller;
+namespace App\Controller\Account;
 
 use App\Entity\Alias;
 use App\Entity\User;
