+++ src/Controller/Settings/ApiTokenController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Controller;
+namespace App\Controller\Settings;
 
 use App\Entity\ApiToken;
 use App\Form\ApiTokenType;
