+++ src/Controller/Settings/DomainController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Controller;
+namespace App\Controller\Settings;
 
 use App\Entity\Domain;
 use App\Entity\User;
