+++ src/Controller/Settings/ReservedNameController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Controller;
+namespace App\Controller\Settings;
 
 use App\Entity\ReservedName;
 use App\Exception\ValidationException;
